arvore pronta, porem falta testar alguns casos, sem leaks
