+++ include/minishell.h
@@ -123,8 +123,24 @@ int		syntax_error(char *msg, t_data *data);
 int		validate_syntax(t_data *data);
 
 // PARSE
-void    parse(t_data *data);
-t_ast   *build_ast(t_token *tokens, t_gc *gc);
+t_token 	*find_operator(t_token *tokens);
+t_token		*find_redir(t_token *tokens);
+t_ast		*parse_subshell(t_token *tokens, t_gc *gc);
+t_ast   	*parse_operator(t_token *tokens, t_token *op, t_gc *gc);
+t_ast   	*parse_cmd(t_token *tokens, t_gc *gc);
+t_ast   	*parse_redir(t_token *tokens, t_token *op, t_gc *gc);
+bool    	is_operator(t_type type);
+bool    	is_redir_bool(t_type type);
+int			is_subshell(t_token *tokens);
+t_token 	*create_token_copy(t_token *src, t_gc *gc);
+t_token		*slice_tokens(t_token *start, t_token *end, t_gc *gc);
+void		handle_error(char *msg, t_data *data);
+t_ast   	*create_node_ast(char *value, t_type type, t_gc *gc);
+int			get_args_len(t_token *tokens);
+char		**extract_args(t_token *tokens, t_gc *gc);
+t_ast   	*build_ast(t_token *tokens, t_gc *gc);
+void    	parse(t_data *data);
+void		print_ast(t_ast *node, int depth);
 
 // GARBAGE COLLECTOR
 void	*gc_malloc(t_gc *gc, size_t size);
