--- src/parse/find_operators.c
@@ -1,2 +1,0 @@
-#include "../../include/minishell.h"
-
